# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Documentation API Server"
        }
    ],
    "paths": {
...
        "/api/v1/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "get all shops. => for system admin and all customers",
                "description": "get all shops",
                "operationId": "get shops",
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "create new shop. => only for shopkeeper",
                "description": "create new shop",
                "operationId": "newShop",
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/shops/{id}": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "get Shop by id in url. => for shopkeeper and system admin",
                "description": "get Shop by id in patch 'url'",
                "operationId": "get shops id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "update shops. => only for shopkeeper and system admin",
                "description": "update shops",
                "operationId": "updateShop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "delete shop. => only for system admin",
                "description": "delete shop",
                "operationId": "deleteShop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "511": {
                        "description": "Network Authentication Required"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in. Just enter the amount of the token.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/api/v1/authorize",
                        "tokenUrl": "http://localhost/api/v1/token",
                        "refreshUrl": "http://localhost/api/v1/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}